FEAT : review

+++ src/main/java/DeliveryFriends/Backend/Domain/Dto/Store/StoreInfoDto.java
@@ -25,7 +25,7 @@ public class StoreInfoDto {
     Boolean packageAvailable;
     Long packageWaitTime;
 
-    Float reviewScore;
+    Long reviewScore;
     Long reviewCount;
     Long orderCount;
